fix hygiene issue with web file system access
